refactor: 기획 변경에 따른 불필요한 코드 삭제 및 변경
