+++ components/Common/index.tsx
@@ -17,5 +17,4 @@ export { default as CommonIconButton } from './IconButton/IconButton';
 export { default as CommonCheckbox } from './Checkbox/Checkbox';
 export { default as CommonFolderButton } from './Tag/FolderButton';
 export { default as CommonLoading } from './Loading/Loading';
-export { default as CommonSelectButton } from './SelectButton/SecondCategorySelect';
 export { default as CommonLogoHeader } from './LogoHeader/LogoHeader';
