--- pages/write/diary.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const diary = () => {
-  return <div>diary</div>;
-};
-
-export default diary;
