--- pages/write/worry.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const worry = () => {
-  return <div>worry</div>;
-};
-
-export default worry;
